Prototyped the feature to hold the property. ( #14 )
Adding the "echonetlite2mqtt.hold" property to the JSON you pass to "properties/set" will hold the property for a period of time.
The format of the "echonetlite2mqtt.hold" property is:

  "{"holdTime":<number> "checkInterval":<number>}"

* holdTime: Time to hold the property (unit: seconds)
* checkInterval: Interval to check properties (unit: seconds) (default: 1)

example:
{"operationStatus":"true", "operationMode":"circulation", "humidifierFunction":"on", "echonetlite2mqtt.hold":{"holdTime":6 "checkInterval":2}}

Check the state once every 2 seconds and hold the state "{"operationStatus":"true", "operationMode":"circulation", "humidifierFunction":"on"}" for 6 seconds.